Moved the typedef file for stylesheet into a types subfolder
