+++ screens/Discovery.tsx
@@ -1,5 +1,5 @@
 import {getDefaultStyleSheet} from "../components/Stylesheet"
-import {StyleSheetI} from "../components/StyleSheetTypes";
+import {StyleSheetI} from "../components/types/StyleSheetTypes";
 import {View, Text, Button} from "react-native";
 import {ThemeContext} from "../components/ThemeContext";
 import React, {useContext} from "react";
