+++ screens/MyRentals.tsx
@@ -3,7 +3,7 @@ import {View, Text} from 'react-native';
 
 import {ThemeContext} from "../components/ThemeContext";
 import {getDefaultStyleSheet} from "../components/Stylesheet";
-import {StyleSheetI} from "../components/StyleSheetTypes";
+import {StyleSheetI} from "../components/types/StyleSheetTypes";
 import {MyRentalsProps} from "./ScreenParams";
 
 export function MyRentals({route, navigation}: MyRentalsProps){
